Removing revit dependencies. Moved UnityLogger.
